move cortex to usecases
