+++ cortex-js/src/infrastructure/dtos/cortex/cortex-operation-successfully.dto.ts
