+++ cortex-js/src/infrastructure/dtos/cortex/start-cortex.dto.ts
