+++ cortex-js/src/infrastructure/exception/model-not-found.exception.ts
