+++ cortex-js/src/usecases/models/models.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { ModelsUsecases } from './models.usecases';
 import { ModelsController } from '../../infrastructure/controllers/models.controller';
 import { DatabaseModule } from 'src/infrastructure/database/database.module';
-import { CortexModule } from 'src/cortex/cortex.module';
+import { CortexModule } from '../cortex/cortex.module';
 
 @Module({
   imports: [DatabaseModule, CortexModule],
